+++ CHANGELOG.md
@@ -47,6 +47,7 @@ See the [full commit history](https://github.com/DFE-Digital/complete-conversion
  - Add new `ProjectTeam` extension method `TeamIsRegionalCaseworkServices`, to identify RCS users 
  - Projects added by you (`/projects/yours/added-by`)
  - Projects completed by you (`/projects/yours/completed`)
+ - Added "order by field" argument to `GetUserWithProjectsQuery`
 
 ### Changed
  - Merged ListAllProjectsByFilter into main ListAllProjects query
@@ -54,6 +55,7 @@ See the [full commit history](https://github.com/DFE-Digital/complete-conversion
  - Allow `ListAllProjectsByFilter` query to handle multiple filters
  - All transfer/conversion projects list use a partial
  - Projects will route to `/project/{projectId}/tasks` from all projects list
+ - Merged `ListAllUsersInTeamWithProjectsQuery` into `ListAllUsersWithProjectsQuery` with filter
 
 ### Fixed
  - Project for user list should show month and year (not day)